SSCSSI-446: removed abstract tests
